Module OCamlR.Interpreter_initialization

Functor used to initialise statically an R interpreter, given initialisation details provided by the provided Env module.

Parameters

Env : Environment

Signature